requirements.txt
Pillow
mss
opencv-python
langchain
langchain-openai
langchain_google_genai
langchain-anthropic
langchain-community
python-dotenv
pyaudio
soundfile
SpeechRecognition
git+https://github.com/openai/whisper.git
